rm some checks
should be fine to pass an empty string to grep or split, as that's more like a "base" or "noop" case. It also prevents errors in queries where the contents are dynamically supplied (a column in another table) and one "poison" empty string causes an error in the entire query.